% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/closest_match.R
\name{closest_match}
\alias{closest_match}
\title{n Closest Match Function}
\usage{
closest_match(df, var_1, var_2, n)
}
\arguments{
\item{df}{data frame that contains all variables in it}

\item{var_1}{numeric number to pivot on}

\item{var_2}{column name in the dataframe to compared vector against var_1}

\item{n}{nth closest variables}
}
\value{
dataframe with n rows that have the closest match
}
\description{
\code{closest_match} This function match n closest distance from a numeric variable.
If the choosen variable rank is the same as others, they are shuffled
and randomly selected.
}
\examples{
closest_match(df, 0.5, "variable", 3)
}
\keyword{closest_match}
