% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lift_func.R
\name{lift_func}
\alias{lift_func}
\title{Lift Calculation Fucntion}
\usage{
lift_func(df, var_lift, var_lift_ind, var_2, n, replacement = FALSE, ...)
}
\arguments{
\item{df}{data frame that contains all variables in it}

\item{var_lift}{column name for the lift variable}

\item{var_lift_ind}{column name for the indicator variable}

\item{var_2}{column name in the dataframe to compared vector against var_1}

\item{n}{nth closest variables}

\item{replacement}{with or without replacement of selection}
}
\value{
data frame with new lift column apended for indicator = 1 rows
}
\description{
This function calculate lift by comparing averages of nth closest pscore
}
\examples{
lift_func(df, "NDS", "target", "pscore", 3)
}
\keyword{closest_match}
