% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/closest_match.R
\name{closest_match}
\alias{closest_match}
\title{n Closest Match Function}
\usage{
closest_match(df, var_1, var_2, n)
}
\arguments{
\item{df}{data frame that contains all variables in it}

\item{var_1}{numeric number to pivot on}

\item{n}{n closest variables}

\item{var_lift}{column name in the dataframe to compared vector against `var_1`}
}
\value{
dataframe with n rows that have the closest match
}
\description{
This function match n closest distance from a numeric variable
}
\examples{
closest_match(df, 0.5, "varaible", 3)
}
\keyword{closest_match}
